feat: 타입 분리, 필요없는 코드 삭제, 구조 변경
