+++ src/types/mypage/mypage.ts
@@ -1,0 +1,7 @@
+export type TSocialPlatform = 'github' | 'kakao' | 'google';
+
+export type TMyProfileValues = {
+  nickname: string;
+  profileImage: string;
+  bannerImage: string;
+};
